Implémentation des rôles pour produit
Corrections diverses

+++ src/main/java/acme/back/service/ProduitService.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 
 import acme.back.db.ProduitDb;
 import acme.back.metier.Produit;
+import acme.front.AuthentificationBean;
 import acme.front.ProduitBean;
 import acme.util.BizException;
 import acme.util.Connexion;
@@ -21,46 +22,52 @@ public class ProduitService {
 
 	
 	public ArrayList<ProduitBean> ProductListBean(AuthentificationBean user) throws BizException {		
-
-		Connexion con = new Connexion();
-		ArrayList<ProduitBean> prodBeans = new ArrayList<ProduitBean>();
-		try {
-			ArrayList<Produit> productList = ProduitDb.getAll(user, con);
-			for (Produit produit : productList) {
-				ProduitBean pb = new ProduitBean();
-				pb.setCodeProduit(produit.getCodeProduit());
-				pb.setLibelleProduit(produit.getLibelleProduit());
-				pb.setPrix(produit.getPrix());
-				pb.setDescription(produit.getDescription());
-				pb.setStimestamp(produit.getStimestamp());
-				prodBeans.add(pb);
+		if (user.hasPermissionToReadProduct()) {
+			Connexion con = new Connexion();
+			ArrayList<ProduitBean> prodBeans = new ArrayList<ProduitBean>();
+			try {
+				ArrayList<Produit> productList = ProduitDb.getAll(user, con);
+				for (Produit produit : productList) {
+					ProduitBean pb = new ProduitBean();
+					pb.setCodeProduit(produit.getCodeProduit());
+					pb.setLibelleProduit(produit.getLibelleProduit());
+					pb.setPrix(produit.getPrix());
+					pb.setDescription(produit.getDescription());
+					pb.setStimestamp(produit.getStimestamp());
+					prodBeans.add(pb);
+				}
+				con.close();
+				return prodBeans;
+	
+			} catch (BizException be) {
+				con.close();
+				throw be;
 			}
-			con.close();
-			return prodBeans;
-
-		} catch (BizException be) {
-			con.close();
-			throw be;
 		}
-
+		else {
+			throw new BizException("Accès refusé pour afficher les produits.");
+	    }
 
 	}
 	
 	public int deleteProduit(AuthentificationBean user, ProduitBean pb) throws BizException {
-		
-		int result;
-		Connexion con = new Connexion();
-		
-		try {
-			con.beginTransaction();
-			result = deleteProduit(user, pb, con);
-			con.endTransaction();
-			return result;
-		} catch (BizException be) {
-			con.rollBack();
-			be.printStackTrace();
-			throw be;
-		}
+		if (user.hasPermissionToDeleteProduct()) {
+			int result;
+			Connexion con = new Connexion();
+			
+			try {
+				con.beginTransaction();
+				result = deleteProduit(user, pb, con);
+				con.endTransaction();
+				return result;
+			} catch (BizException be) {
+				con.rollBack();
+				be.printStackTrace();
+				throw be;
+			}
+		}else {
+			throw new BizException("Accès refusé pour supprimer un produit.");
+	    }
 	}
 	
 	public int deleteProduit(AuthentificationBean user, ProduitBean pb, Connexion con) throws BizException {
@@ -76,38 +83,44 @@ public class ProduitService {
 			}
 			
 		}else {
-	        System.out.println("Accès refusé pour supprimer un produit.");
+			throw new BizException("Accès refusé pour supprimer un produit.");
 	    }
 	}
 	
 	public int updateProduit(AuthentificationBean user, ProduitBean pb) throws BizException {
+		if (user.hasPermissionToUpdateProduct()) {
+			int result;
+			Connexion con = new Connexion();			
+			try {
+				con.beginTransaction();
+				result = updateProduit(user, pb, con);
+				con.endTransaction();
+				return result;
+			} catch (BizException be) {
+				con.rollBack();
+				be.printStackTrace();
+				throw be;
+			}
+		}else {
+			throw new BizException("Accès refusé pour mettre à jour un produit.");
+	    }
 		
-		int result;
-		Connexion con = new Connexion();
-		
-		try {
-			con.beginTransaction();
-			result = updateProduit(user, pb, con);
-			con.endTransaction();
-			return result;
-		} catch (BizException be) {
-			con.rollBack();
-			be.printStackTrace();
-			throw be;
-		}
 	}
 	
 	public int updateProduit(AuthentificationBean user, ProduitBean pb, Connexion con) throws BizException {
-		
-		int result = 0;
-		Produit p = produitBeanToProduit(pb);
-		try {
-			System.out.println("modification produit = " + ProduitDb.updateByKey(user, con, p));
-			return result;
-		} catch (BizException be) {
-			be.printStackTrace();
-			throw be;
-		}
+		if (user.hasPermissionToUpdateProduct()) {
+			int result = 0;
+			Produit p = produitBeanToProduit(pb);
+			try {
+				System.out.println("modification produit = " + ProduitDb.updateByKey(user, con, p));
+				return result;
+			} catch (BizException be) {
+				be.printStackTrace();
+				throw be;
+			}
+		}else {
+			throw new BizException("Accès refusé pour mettre à jour un produit.");
+	    }
 	}
 	
 	public int insertProduit(AuthentificationBean user, ProduitBean pb) throws BizException {
@@ -126,14 +139,14 @@ public class ProduitService {
 				throw be;
 			}
 		}else {
-            System.out.println("Accès refusé pour créer un produit.");
+			throw new BizException("Accès refusé pour créer un produit.");
         }
 	}
 	
 	public int insertProduit(AuthentificationBean user, ProduitBean pb, Connexion con) throws BizException {
 		if (user.hasPermissionToCreateProduct()) {
 			int result = 0;
-			Produit p = produitBeanToProduit(user, pb);
+			Produit p = produitBeanToProduit(pb);
 			try {
 				System.out.println("ajout produit = " + ProduitDb.insert(user, con, p));
 				return result;
@@ -142,7 +155,7 @@ public class ProduitService {
 				throw be;
 			}
 		}else {
-            System.out.println("Accès refusé pour créer un produit.");
+			throw new BizException("Accès refusé pour créer un produit.");
         }
 	}
 	
@@ -162,7 +175,7 @@ public class ProduitService {
 				throw be;
 			}
 		}else {
-            System.out.println("Accès refusé pour récupérer un produit.");
+			throw new BizException("Accès refusé pour afficher un produit.");
         }
 	}
 	
@@ -178,7 +191,7 @@ public class ProduitService {
 				throw be;
 			}
 		}else {
-            System.out.println("Accès refusé pour récupérer un produit.");
+			throw new BizException("Accès refusé pour afficher un produit.");
         }
 	}
